build(ci): Switch npm dependencies to pnpm in Github Actions
This commit replaces `npm` with `pnpm` in the Github Actions workflows. `pnpm` is used due to its efficient handling of dependencies, reducing the space and installation time required. In each workflow, `actions/setup-node@v2` is updated to `actions/setup@v3`, the install command is switched to `pnpm install --frozen-lockfile` and all the npm run commands are switched to pnpm equivalents. This enhances overall workflow performance.